added labels to knative event source CRD response
